move model loading to separate module
